EXtract common queueProcessor
