+++ web/apps/cast/src/services/wasm/ffmpeg.ts
@@ -1,9 +1,9 @@
 import { addLogLine } from "@ente/shared/logging";
 import { promiseWithTimeout } from "@ente/shared/promise";
 import { logError } from "@ente/shared/sentry";
+import QueueProcessor from "@ente/shared/utils/queueProcessor";
 import { generateTempName } from "@ente/shared/utils/temp";
 import { createFFmpeg, FFmpeg } from "ffmpeg-wasm";
-import QueueProcessor from "services/queueProcessor";
 import { getUint8ArrayView } from "services/readerService";
 
 const INPUT_PATH_PLACEHOLDER = "INPUT";
