+++ web/apps/cast/src/services/wasmHeicConverter/wasmHEICConverterService.ts
@@ -2,8 +2,8 @@ import { CustomError } from "@ente/shared/error";
 import { addLogLine } from "@ente/shared/logging";
 import { retryAsyncFunction } from "@ente/shared/promise";
 import { logError } from "@ente/shared/sentry";
+import QueueProcessor from "@ente/shared/utils/queueProcessor";
 import { convertBytesToHumanReadable } from "@ente/shared/utils/size";
-import QueueProcessor from "services/queueProcessor";
 import { getDedicatedConvertWorker } from "utils/comlink/ComlinkConvertWorker";
 import { ComlinkWorker } from "utils/comlink/comlinkWorker";
 import { DedicatedConvertWorker } from "worker/convert.worker";
