+++ web/apps/photos/src/services/wasm/ffmpeg.ts
@@ -2,7 +2,7 @@ import { addLogLine } from "@ente/shared/logging";
 import { logError } from "@ente/shared/sentry";
 import { generateTempName } from "@ente/shared/utils/temp";
 import { createFFmpeg, FFmpeg } from "ffmpeg-wasm";
-import QueueProcessor from "services/queueProcessor";
+import QueueProcessor from "@ente/shared/utils/queueProcessor";
 import { getUint8ArrayView } from "services/readerService";
 import { promiseWithTimeout } from "utils/common";
